Add a basic Prover system
This patch adds a basic prover system to prove (haha!) that we can verify data exists in an IPFS tree by simply having a root_hash (e.g. block hash), path (known in advance), data (known in advance) and a list of proof (returned from this protocol). Any party with the first three elements and being given the fourth element (proof) can verify that the data exists at the path as expected, then though that user has no knowledge of the data in the IPFS tree itself.

This is important because a blockchain contract might say "I will give you 3Eth if you set the value of the blockchain at /accounts/my_account/value to 3." This allows that blockchain contract to verify that value and then "accept" that the contract has been paid in full. The side-chain can add this value to the IPFS tree, post a new blockhash, and call this contract to fulfill the contract, providing the proof itself. If there's no proof provided, the original caller can rescind his/her original Eth and void the contract.

Proof? What's the proof. The proof here is the protobuf-encoded nodes from the leaf to the root. The verifier can at every step ensure that the nodes match his/her expectations from leaf (the data should be `3`), to the leaf's parent (there should exist a link at `value` with the sha256 sum of the leaf) to it's parent (there should exist a valid at `my_account` with the sha256 of the previous link node) all the way to the top. If at any point we fail to match the sha256 link at the right sub-path, we fail.

This algorithm depends on decoding protobufs, but it's actually quite feasible since we know the structure in advance and it's fairly easy to read data out of protobufs if you know the structure. We also need to decode varints (pretty easy) and perform sha256 sums (easy to do in the Ethereum blockchain). Thus, it sounds quite plausible that we can encode this entire proof system in an Ethereum Bridge Contract which will allow anyone to on or off-rail from the sidechain.