Build block serializer in IPFS
Previously, we were encoding blocks in Protobuf and pushing that data to IPFS. This makes it difficult to partially load block data (e.g. to just pull the final state) or to explore the blockchain data. Instead, we now opt to store the serialized block as a set of unix-like files in IPFS. This patch completes the transition to this new system. As a bonus, we start to add proper test cases for more modules.