[WIP] Begin to add Persistence model
Currently, every time the protocol starts, it begins with a brand-new blockchain from scratch. We should instead opt to store the current block hash whenever we mine a new block and then push that blockhash to IPNS so that peers can also start from that blockhash. This patch begins to implement that strategy.