Add better key management and separate minter from publisher.
This patch makes some changes to our persistence system to make it more stable. Specifically, publishing is a very slow process (takes currently around 60 seconds), so we need to change the direction of our publishing. That is, instead of the minter occausionally publishing new blocks, we have a new Publisher which every so often asks the Minter to mint a completed block and then the Publisher goes off to publish the block. During this time, the Minter now can add to the next block and it will mint that one as soon as the Publisher asks for it. This flow is significantly more stable.

Additionally, we now have keys managed by IPFS and allow in the config the user to pass a given key. This means that a user can use whatever key he/she wants to publish, so long as it's been loaded into the IPFS keychain.