Add a Leader / Follower System
This patch begins to add a leader / follower system. If a given node is a leader, it will mint and publish blocks to IPFS. If a node is a follower, instead it will read blocks from IPFS, verify those blocks, and if they match expectations, will add them to be read.