Add a Link System
This patch adds a simple link system that let's us now store all data via IPFS file system links (whereas we used to just stoe the hash of other IPFS trees, we now store them as proper links). This seemingless small change makes it very easy to view and understand the entire block just from looking at IPFS. We currently just label such links as "<name>_link" so that we know not to fully load them  (and instead, to load that data on demand as needed).